[Cordova] Add cordova 4.x support
- Move some usecases to demo express
- Fail Analyse:XWALK-2233/2648

BUG=https://crosswalk-project.org/jira/browse/XWALK-6056
BUG=https://crosswalk-project.org/jira/browse/XWALK-5864
