+++ VERSION
@@ -1,6 +1,6 @@
 {
     "app-version": "v1.0",
-    "main-version": "v1.0",
+    "main-version": "16.45.421.19",
     "release-version": "1",
-    "crosswalk-branch": "beta"
+    "crosswalk-branch": "stable"
 }
