+++ samples-cordova/AdMob4.x/COPYING
@@ -1,0 +1,7 @@
+This test comes from
+https://github.com/floatinghotpot/cordova-admob-pro/tree/master/test
+with the modification:
+Add bootstrap framework.
+
+The test is dual-licensed under the MIT License:
+https://github.com/floatinghotpot/cordova-admob-pro/blob/master/LICENSE
