+++ samples-cordova/AdMob4.x/README.md
@@ -1,0 +1,3 @@
+## Usecase Design
+
+This sample demonstrates Cordova AdMob plugin can create, remove, show, hide, interstitial the Ad.
