+++ samples-cordova/AdMob4.x/index.html
@@ -1,0 +1,113 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="utf-8" />
+        <meta name="format-detection" content="telephone=no" />
+        <meta name="viewport" content="user-scalable=no, initial-scale=1, maximum-scale=1, minimum-scale=1, width=device-width" />
+        <title>AdMob</title>
+        <link rel="stylesheet" type="text/css" href="../../css/bootstrap.css">
+        <link rel="stylesheet" type="text/css" href="../../css/main.css">
+        <script src="../../js/jquery-2.1.3.min.js"></script>
+        <script src="../../js/bootstrap.min.js"></script>
+        <script src="../../js/common.js"></script>
+        <script src="../../js/tests.js"></script>
+        <script src="../../main.js"></script>
+        <script src="../../cordova.js"></script>
+        <style type="text/css">
+            html, body { width:100%; height:100%; margin:0; padding:0; overflow:hidden; background-color:white; }
+            div#fullpage { width:100%; height:100%; margin:0; padding-top: 60px; border:0px solid red; text-align:center; vertical-align:middle; }
+            button { font-size: 22px; }
+        </style>
+    </head>
+    <body onload="onDocLoad()" onresize="onResize()">
+    <script>
+    function onDocLoad() {
+        if(( /(ipad|iphone|ipod|android|windows phone)/i.test(navigator.userAgent) )) {
+            document.addEventListener('deviceready', initApp, false);
+        } else {
+            initApp();
+        }
+    }
+    
+	var admobid = {};
+	if( /(android)/i.test(navigator.userAgent) ) { 
+		admobid = { // for Android
+			banner: 'ca-app-pub-6869992474017983/9375997553',
+			interstitial: 'ca-app-pub-6869992474017983/1657046752'
+		};
+	} else if(/(ipod|iphone|ipad)/i.test(navigator.userAgent)) {
+		admobid = { // for iOS
+			banner: 'ca-app-pub-6869992474017983/4806197152',
+			interstitial: 'ca-app-pub-6869992474017983/7563979554'
+		};
+	} else {
+		admobid = { // for Windows Phone
+			banner: 'ca-app-pub-6869992474017983/8878394753',
+			interstitial: 'ca-app-pub-6869992474017983/1355127956'
+		};
+	}
+        
+    function initApp() {
+		if (typeof window.AdMob === 'undefined') { alert( 'admob plugin not ready' ); return; }
+		initAd();
+        // display the banner at startup
+        AdMob.createBanner( { adId:admobid.banner, overlap:flase } );
+    }
+    function initAd(){
+        var defaultOptions = {
+            // adSize: 'SMART_BANNER',
+            // width: integer, // valid when set adSize 'CUSTOM'
+            // height: integer, // valid when set adSize 'CUSTOM'
+            position: AdMob.AD_POSITION.BOTTOM_CENTER,
+            // offsetTopBar: false, // avoid overlapped by status bar, for iOS7+
+            bgColor: 'black', // color name, or '#RRGGBB'
+            // x: integer,		// valid when set position to 0 / POS_XY
+            // y: integer,		// valid when set position to 0 / POS_XY
+            isTesting: true, // set to true, to receiving test ad for testing purpose
+            // autoShow: true // auto show interstitial ad when loaded, set to false if prepare/show
+        };
+        AdMob.setOptions( defaultOptions );
+        registerAdEvents();
+    }
+    // optional, in case respond to events or handle error
+    function registerAdEvents() {
+        // new events, with variable to differentiate: adNetwork, adType, adEvent
+        document.addEventListener('onAdFailLoad', function(data){ 
+        	alert('error: ' + data.error + 
+        			', reason: ' + data.reason + 
+        			', adNetwork:' + data.adNetwork + 
+        			', adType:' + data.adType + 
+        			', adEvent:' + data.adEvent); // adType: 'banner' or 'interstitial'
+        });
+        document.addEventListener('onAdLoaded', function(data){});
+        document.addEventListener('onAdPresent', function(data){});
+        document.addEventListener('onAdLeaveApp', function(data){});
+        document.addEventListener('onAdDismiss', function(data){});
+    }
+    function onResize() {
+        var msg = 'web view: ' + window.innerWidth + ' x ' + window.innerHeight;
+        document.getElementById('sizeinfo').innerHTML = msg;
+    }
+    </script>
+    <div id="header">
+      <h3 id="main_page_title"></h3>
+    </div>
+    <div class="content">
+      <div id="fullpage">
+        <h4>This sample demonstrates the Ad can be created, removed, shown, hidden, interstitial.</h4>
+        <p><button onclick="AdMob.createBanner( { adId:admobid.banner, autoShow:true } );">create Ad</button> <button onclick="AdMob.removeBanner();">remove Ad</button></p>
+        <p><button onclick="AdMob.showBanner(AdMob.AD_POSITION.BOTTOM_CENTER);">show Ad</button> <button onclick="AdMob.hideBanner();">hide Ad</button></p>
+        <p><button onclick="AdMob.prepareInterstitial({adId:admobid.interstitial, autoShow:true});">create Interstitial Ad</button></p>
+        <p><button onclick="AdMob.showInterstitial();">show Interstitial Ad</button></p>
+        <div id="sizeinfo">width * height</div>
+        <div>Try rotate screen to test the orientation change</div>
+      </div>
+    </div>
+    <div class="footer">
+      <div id="footer"></div>
+    </div>
+    <div class="modal fade" id="popup_info">
+      <p>AdMob features: createBanner, removeBanner, showBanner, hideBanner, prepareInterstitial, showInterstitial method.</p>
+    </div>
+    </body>
+</html>
