+++ samples-cordova/InAppBrowser/REAMDE.md
@@ -1,0 +1,12 @@
+## Usecase Design
+
+This sample demonstrates CordovaWebView feature basic functionalities, include:
+
+* Open remote PDF and local PDF in window by default 
+* Open remote PDF and local PDF in current window
+* Share the cookie of main page
+
+This usecase covers following method:
+
+* Window interface: open() method
+* Document interface: cookie attribute
