+++ samples-cordova/InAppBrowser/index.html
@@ -1,0 +1,63 @@
+<!DOCTYPE html>
+<!--
+Copyright (c) 2014 Intel Corporation.
+
+Redistribution and use in source and binary forms, with or without modification,
+are permitted provided that the following conditions are met:
+
+* Redistributions of works must retain the original copyright notice, this list
+  of conditions and the following disclaimer.
+* Redistributions in binary form must reproduce the original copyright notice,
+  this list of conditions and the following disclaimer in the documentation
+  and/or other materials provided with the distribution.
+* Neither the name of Intel Corporation nor the names of its contributors
+  may be used to endorse or promote products derived from this work without
+  specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
+AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
+INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
+OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
+EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+Authors:
+        Wang, Hongjuan <hongjuanx.wang@intel.com>
+        Li, Hao <haox.li@intel.com>
+
+-->
+
+<meta charset="utf-8" />
+<meta name="viewport" content="initial-scale=1.0, maximum-scale=1.0, width=device-width" />
+<link rel="stylesheet" type="text/css" href="../../css/bootstrap.css">
+<link rel="stylesheet" type="text/css" href="../../css/main.css">
+<script src="../../js/jquery-2.1.3.min.js"></script>
+<script src="../../js/bootstrap.min.js"></script>
+<script src="../../js/common.js"></script>
+<script src="../../js/tests.js"></script>
+<script src="../../cordova.js"></script>
+<script src="js/main.js"></script>
+
+<body onload="init();">
+  <div id="header">
+    <h3 id="main_page_title"></h3>
+  </div>
+  <div class="content">
+    <h4>This sample demonstrates the remote pdf, local pdf and cookie page can be opened.</h4>
+    <button onclick="doOpen('https://github.com/crosswalk-project/crosswalk-test-suite/tree/master/doc');" class="btn btn-default btn-lg btn-block">Open remote PDF(target=Default)</button>
+    <button onclick="doOpen('local.pdf');" class="btn btn-default btn-lg btn-block">Open local PDF(target=Default)</button>
+    <button onclick="doOpen('https://github.com/crosswalk-project/crosswalk-test-suite/tree/master/doc', '_self');" class="btn btn-default btn-lg btn-block">Open remote PDF(target=_self)</button>
+    <button onclick="doOpen('local.pdf', '_self');" class="btn btn-default btn-lg btn-block">Open local PDF(target=_self)</button>
+    <button onclick="doOpen('get-cookie.html', '_blank');" class="btn btn-default btn-lg btn-block">Open Cookie Page</button>
+  </div>
+  <div class="footer">
+    <div id="footer"></div>
+  </div>
+  <div class="modal fade" id="popup_info">
+    <p>InAppBrowser features: open method, cookie attribute.</p>
+  </div>
+</body>
