+++ samples-cordova/Notification/REAMDE.md
@@ -1,0 +1,12 @@
+## Usecase Design
+
+This sample demonstrates Notification feature basic functionalities, include:
+
+* Beep notification
+* Vibration notification
+* Confirm dialog
+* Prompt dialog
+
+This usecase covers following interface and methods:
+
+* Notification interface: beep(), vibrate(), confirm(), prompt() methods
